SCP-4111 Split utilities into separate modules.
